+++ popkorn-example/src/main/java/cc/popkorn/example/ExampleCompat.java
@@ -1,15 +1,7 @@
 package cc.popkorn.example;
 
 import cc.popkorn.PopKornCompat;
-import cc.popkorn.example.model.D10;
-import cc.popkorn.example.model.R1i;
-import cc.popkorn.example.model.R2i;
-import cc.popkorn.example.model.R3i;
-import cc.popkorn.example.model.R4i;
-import cc.popkorn.example.model.R5i;
-import cc.popkorn.example.model.R6i;
-import cc.popkorn.example.model.R7i;
-import cc.popkorn.example.model.R8i;
+import cc.popkorn.example.model.*;
 
 public class ExampleCompat {
 
@@ -36,6 +28,7 @@ public class ExampleCompat {
         PopKornCompat.inject(R8i.class,"env2");
         PopKornCompat.inject(R8i.class,"env3");
         PopKornCompat.inject(R8i.class,"env4");
+        PopKornCompat.inject(R9i.class);
 
         PopKornCompat.removeInjectable(D10.class);
         PopKornCompat.reset();