+++ popkorn-example/src/main/kotlin/cc/popkorn/example/model/Ri.kt
@@ -16,3 +16,5 @@ interface R7i
 
 interface R8i
 
+interface R9i
+